requirements.txt
# Web 框架
Flask==3.0.0
Flask-CORS==4.0.0

# 数据库
pymongo==4.6.0
dnspython==2.4.2

# 定时任务
APScheduler==3.10.4

# 爬虫相关
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
chardet==5.2.0
selenium==4.15.0

# 数据处理
pandas==2.1.3
openpyxl==3.1.2

# 配置和环境
python-dotenv==1.0.0

# API 文档（可选）
flasgger==0.9.7.1
